% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_display.R
\name{b_disp_cal_data}
\alias{b_disp_cal_data}
\title{Display calibration data}
\usage{
b_disp_cal_data(cal_data)
}
\arguments{
\item{cal_data}{numeric matrix as returned by \code{\link{b_read_cal_data}}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Prints the calibration data matrix to the console.
}
