% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_mc.R
\name{b_disp_cal_data_mc}
\alias{b_disp_cal_data_mc}
\title{Display calibration data summary from Monte Carlo samples}
\usage{
b_disp_cal_data_mc(cal_samples)
}
\arguments{
\item{cal_samples}{array as returned by
\code{\link{b_sample_cal_data_mc}}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Display calibration data summary from Monte Carlo samples
}
\keyword{internal}
