% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_mc.R
\name{b_disp_cal_results_mc}
\alias{b_disp_cal_results_mc}
\title{Display calibration results from Monte Carlo samples}
\usage{
b_disp_cal_results_mc(b_samples)
}
\arguments{
\item{b_samples}{matrix as returned by \code{\link{b_least_mc}}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Display calibration results from Monte Carlo samples
}
