% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_mc.R
\name{b_disp_meas_results_mc}
\alias{b_disp_meas_results_mc}
\title{Display measurement results from Monte Carlo samples}
\usage{
b_disp_meas_results_mc(x_samples, meas_samples)
}
\arguments{
\item{x_samples}{matrix as returned by \code{\link{b_eval_mc}}.}

\item{meas_samples}{matrix as returned by
\code{\link{b_sample_meas_data_mc}}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Display measurement results from Monte Carlo samples
}
\keyword{internal}
