% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_mc.R
\name{b_eval_mc}
\alias{b_eval_mc}
\title{Evaluate calibration function for each Monte Carlo sample}
\usage{
b_eval_mc(meas_samples, b_samples, func)
}
\arguments{
\item{meas_samples}{matrix as returned by
\code{\link{b_sample_meas_data_mc}}.}

\item{b_samples}{matrix as returned by \code{\link{b_least_mc}}.}

\item{func}{calibration function.}
}
\value{
A matrix of shape \code{[nsamples, m]} of assigned values.
}
\description{
For each sample, applies \code{func} with the corresponding row of
\code{b_samples} to the corresponding row of \code{meas_samples}.
}
