% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_read.R
\name{b_example_data_path}
\alias{b_example_data_path}
\title{Path to a package example data file}
\usage{
b_example_data_path(example, type)
}
\arguments{
\item{example}{integer, 1, 2, or 3.}

\item{type}{character, \code{"cal"} for calibration data or
\code{"meas"} for measurement data.}
}
\value{
A character string giving the full path to the file.
}
\description{
Returns the full path to one of the six example data files from Annex B
of ISO 6143:2001 that are shipped with the package.
}
