% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_mc.R
\name{b_least_mc}
\alias{b_least_mc}
\title{Fit calibration function to each Monte Carlo sample}
\usage{
b_least_mc(cal_samples, func)
}
\arguments{
\item{cal_samples}{array as returned by
\code{\link{b_sample_cal_data_mc}}.}

\item{func}{calibration function.}
}
\value{
A matrix of shape \code{[nsamples, nb]} where \code{nb} is the
  number of coefficients.
}
\description{
For each of the \code{nsamples} calibration trials, runs
Levenberg-Marquardt (without analytic Jacobian for speed) and collects the
fitted coefficients.
}
