% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_mc.R
\name{b_mean_cov_mc}
\alias{b_mean_cov_mc}
\title{Mean and covariance of Monte Carlo samples}
\usage{
b_mean_cov_mc(samples)
}
\arguments{
\item{samples}{numeric matrix where each row is one sample.}
}
\value{
A list with elements \code{mean} and \code{cov}.
}
\description{
Mean and covariance of Monte Carlo samples
}
