% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_mc.R
\name{b_test_mc}
\alias{b_test_mc}
\title{Full Monte Carlo calibration and evaluation with printed output}
\usage{
b_test_mc(cal_data, meas_data, func, nsamples = 10000L)
}
\arguments{
\item{cal_data}{numeric matrix as returned by \code{\link{b_read_cal_data}}.}

\item{meas_data}{numeric matrix as returned by
\code{\link{b_read_meas_data}}.}

\item{func}{calibration function.}

\item{nsamples}{integer number of Monte Carlo samples (default 10000).}
}
\value{
Invisibly a list with elements \code{b_samples} and
  \code{x_samples}.
}
\description{
Full Monte Carlo calibration and evaluation with printed output
}
