% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_names.R
\name{componentName}
\alias{componentName}
\title{Return the name of a gas component by index}
\usage{
componentName(index)
}
\arguments{
\item{index}{Integer 1–60.}
}
\value{
Character string with the English component name.
}
\description{
Return the name of a gas component by index
}
