% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ISO6143.2001-package}
\alias{ISO6143.2001}
\alias{ISO6143.2001-package}
\title{ISO6143.2001: Calibration of Gas Analysers per ISO 6143:2001}
\description{
Implements the B LEAST calibration algorithm of ISO 6143:2001.
See \code{\link{b_least}} for the main fitting function and
\code{\link{b_eval}} for evaluating fitted functions at measurement points.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RuedigerForster/ISO_6143}
  \item Report bugs at \url{https://github.com/RuedigerForster/ISO_6143/issues}
}

}
\author{
\strong{Maintainer}: Rüdiger Forster \email{meticulous.measurements@gmail.com}

Authors:
\itemize{
  \item Wolfram Bremser (Author of the original B_LEAST program (BAM))
}

}
\keyword{internal}
