% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iso6976_calc}
\alias{iso6976_calc}
\title{Calculate all ISO 6976:2016 properties and their uncertainties}
\usage{
iso6976_calc(x, u_x, r_x, t1, t2, p2, k = 1)
}
\arguments{
\item{x}{Numeric vector of length 60: mole fractions [mol/mol]}

\item{u_x}{Numeric vector of length 60: standard uncertainties}

\item{r_x}{Numeric matrix 60x60: correlation matrix}

\item{t1}{Combustion temperature [°C]: 0, 15, 15.55, 20, or 25}

\item{t2}{Volume reference temperature [°C]: 0, 15, 15.55, or 20}

\item{p2}{Reference pressure [kPa]: 90–110}

\item{k}{Coverage factor (default 1)}
}
\value{
Named list of properties and uncertainties
}
\description{
Calculate all ISO 6976:2016 properties and their uncertainties
}
